Fix Werkzeug 0.15 incompatibility
Wekzeug 0.15 is incompatibility Flask versions less than v1.0.0.
Specifically the test client will fail if Werkzeug version 0.15 is used
with e.g. Flask 0.11. Specifying the minimum Werkzeug as a requirements
ensure that we always install the lowest Werkzeug version with the
lowest Flask version, and highst with highest (due to
requirements-builder).